chore: merge duplicated initPlugins modules
