+++ packages/compat/webpack/src/ProgressPlugin/ProgressPlugin.ts
