+++ packages/compat/webpack/src/ProgressPlugin/helpers/bus.ts
@@ -1,6 +1,6 @@
 import { Console } from 'console';
-import patchConsole from '../../../../compiled/patch-console';
-import cliTruncate from '../../../../compiled/cli-truncate';
+import patchConsole from '../../../compiled/patch-console/build';
+import cliTruncate from '../../../compiled/cli-truncate';
 import type { Props } from './type';
 import { FULL_WIDTH, renderBar } from './bar';
 import { create } from './log';
