+++ packages/compat/webpack/src/ProgressPlugin/helpers/index.ts
