+++ packages/compat/webpack/src/ProgressPlugin/helpers/nonTty.ts
