+++ packages/compat/webpack/src/ProgressPlugin/helpers/percentage.ts
