+++ packages/compat/webpack/src/ProgressPlugin/helpers/type.ts
