+++ packages/compat/webpack/src/ProgressPlugin/helpers/utils.ts
