+++ packages/compat/webpack/src/core/createCompiler.ts
@@ -7,7 +7,8 @@ import {
   type Rspack,
   type RspackConfig,
 } from '@rsbuild/shared';
-import type { Context, WebpackConfig } from '../types';
+import type { Context } from '@rsbuild/core/rspack-provider';
+import type { WebpackConfig } from '../types';
 import { initConfigs, InitConfigsOptions } from './initConfigs';
 import type { Compiler, MultiCompiler } from 'webpack';
 import { getDevMiddleware } from './devMiddleware';
