+++ packages/compat/webpack/src/core/webpackConfig.ts
@@ -10,10 +10,11 @@ import {
   type ModifyWebpackChainUtils,
   type ModifyWebpackConfigUtils,
 } from '@rsbuild/shared';
+import type { Context } from '@rsbuild/core/rspack-provider';
 import { getCompiledPath } from '../shared';
 import type { RuleSetRule, WebpackPluginInstance } from 'webpack';
 
-import type { Context, WebpackConfig } from '../types';
+import type { WebpackConfig } from '../types';
 
 async function modifyWebpackChain(
   context: Context,
