+++ packages/compat/webpack/src/index.ts
@@ -1,13 +1,6 @@
 export { webpackProvider } from './provider';
 export { webpackBuild } from './core/build';
 export type {
-  LessLoaderOptions,
-  SassLoaderOptions,
-  PostCSSLoaderOptions,
-  CSSLoaderOptions,
-  StyleLoaderOptions,
-} from '@rsbuild/shared';
-export type {
   // Third Party Types
   webpack,
   WebpackChain,
