--- packages/compat/webpack/src/shared/fs.ts
@@ -1,15 +1,0 @@
-import fs from 'fs';
-import { join } from 'path';
-import {
-  getSharedPkgCompiledPath,
-  SharedCompiledPkgNames,
-} from '@rsbuild/shared';
-
-export const getCompiledPath = (packageName: string) => {
-  const providerCompilerPath = join(__dirname, '../../compiled', packageName);
-  if (fs.existsSync(providerCompilerPath)) {
-    return providerCompilerPath;
-  } else {
-    return getSharedPkgCompiledPath(packageName as SharedCompiledPkgNames);
-  }
-};
