--- packages/compat/webpack/src/shared/index.ts
@@ -1,1 +1,0 @@
-export * from './fs';
