+++ packages/core/src/rspack-provider/index.ts
@@ -5,6 +5,9 @@ export {
   createPublicContext,
   createContextByConfig,
 } from './core/createContext';
+export { initPlugins } from '@rsbuild/shared';
 export { initHooks, type Hooks } from './core/initHooks';
 export { withDefaultConfig } from './config';
 export { initRsbuildConfig } from './core/initConfigs';
+export { getPluginAPI } from './core/initPlugins';
+export type { Context } from '../types';
